requirements.txt
python-dotenv
twitchio
requests